+++ WireUtils.h
@@ -2,7 +2,7 @@
 #define _WIRE_UTILS_H_
 #include <Arduino.h>
 
-byte ProbeAddress(byte address);
-byte WriteRegisters(byte address, byte regAddr, int byteCount, byte *buffer);
-byte ReadRegisters(byte address, byte regAddr, int byteCount, byte *buffer);
+byte ProbeAddress(uint8_t address);
+byte WriteRegisters(uint8_t address, uint8_t regAddr, uint8_t byteCount, byte *buffer);
+byte ReadRegisters(uint8_t address, uint8_t regAddr, uint8_t byteCount, byte *buffer);
 #endif
